requirements.txt
colorama
requests
python-nmap
dnspython
beautifulsoup4
lxml
html5lib
wheel
